requirements.txt
Flask==3.0.0
PyPDF2==3.0.1
python-docx==0.8.11
gunicorn==21.2.0

